EventRepository.java

+++ Calendar/Calendar/src/main/java/com/example/Calendar/Database/EventRepository.java
@@ -1,9 +1,9 @@
 package com.example.Calendar.Database;
 
-import jdk.jfr.Event;
+import com.example.Calendar.Models.Event;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EventRepository extends JpaRepository<Event, Long> {
+public interface EventRepository extends JpaRepository<Event, Integer> {
 }
